jrcu: add new stat to /sys/kernel/debug/rcu/rcudata
jrcu: display #passes in /sys/kernel/debug/rcu/rcudata.

Sometimes, when things are not working right, you cannot
tell if that is because the daemon is not running at all,
or because it is running, but on every pass it decides
to do no work.  With this change those two states can
now be distinguished.

Signed-off-by: Joe Korty <joe.korty@ccur.com>